simplify API
